GCW-3417 Rename GoodCityApiV2Client to client
