+++ src/lib/services/AuthenticationService/AuthenticationService.ts
@@ -1,4 +1,4 @@
-import GoodCityApiV2Client from "lib/client/GoodCityApiV2Client/GoodCityApiV2Client";
+import client from "lib/client/client";
 
 interface SendPinBody {
   mobile: string;
@@ -7,7 +7,7 @@ interface SendPinResponse {
   otp_auth_key: string;
 }
 function sendPin(body: SendPinBody): Promise<SendPinResponse> {
-  return GoodCityApiV2Client("auth/send_pin", body);
+  return client("auth/send_pin", body);
 }
 
 interface VerifyBody {
@@ -18,7 +18,7 @@ interface VerifyResponse {
   jwt_token: string;
 }
 function verify(body: VerifyBody): Promise<VerifyResponse> {
-  return GoodCityApiV2Client("auth/verify", body);
+  return client("auth/verify", body);
 }
 
 const AuthenticationService = {
